requirements.txt
# Core
python-dotenv
requests
tqdm

# Data / plotting (only if you actually use them)
numpy
pandas
matplotlib


langchain-google-genai
langchain


